slave_deploy/requirements.txt
Adafruit_DHT
flask
uwsgi
RPi.GPIO
